requirements/requirements.txt
asgiref==3.6.0
django==4.2.0
python-dotenv==0.21.1
sqlparse==0.4.3
